Reorg track validators performance cron
