+++ packages/brain/src/modules/cron/index.ts
@@ -1,4 +1,7 @@
 export { CronJob } from "./cron.js";
 export { reloadValidators } from "./reloadValidators/index.js";
 export { sendProofsOfValidation } from "./sendProofsOfValidation/index.js";
-export { trackValidatorsPerformance } from "./trackValidatorsPerformance/index.js";
+export {
+  trackValidatorsPerformanceCron,
+  startWithinTenFirstPercentageOfEpoch
+} from "./trackValidatorsPerformance/index.js";
